Moved type registration to its own plugin to clean up a bit.
