+++ src/game_state/game_state_plugin.rs
@@ -22,6 +22,7 @@ use crate::towers::systems::{shoot_alien_system, tower_has_alien_in_range_scorer
 use crate::ui::spawn_ui::{add_health_bar, AddHealthBar, fellow_system, GotoState};
 use crate::ui::ui_plugin::UiPlugin;
 use crate::generate_mesh::MeshPlugin;
+use crate::inspection::inspector::InspectorPlugin;
 use crate::playground::PlaygroundPlugin;
 
 pub struct GamePlugin;
@@ -48,7 +49,7 @@ impl Plugin for GamePlugin {
                 ScoreKeeperPlugin,
                 // VideoGlitchPlugin,
                 GamepadPlugin,
-                // InspectorPlugin,
+                InspectorPlugin,
             ))
             .add_plugins((
                 MeshPlugin,
